Packets: Correct packet sizes in all Msg packets

+++ src/world/Server/Packets/MsgTalentWipeConfirm.h
@@ -29,10 +29,7 @@ namespace AscEmu::Packets
         }
 
     protected:
-        size_t expectedSize() const override
-        {
-            return 8 + 4;
-        }
+        size_t expectedSize() const override { return static_cast<size_t>(8 + 4); }
 
         bool internalSerialise(WorldPacket& packet) override
         {
